[shuffle] break out Config into shared module
