+++ shuffle/cli/src/node.rs
@@ -1,10 +1,10 @@
 // Copyright (c) The Diem Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::new::Config;
+use crate::shared::read_config;
 use anyhow::Result;
 use diem_config::config::NodeConfig;
-use std::{fs, path::Path};
+use std::path::Path;
 
 pub fn handle(project_path: &Path) -> Result<()> {
     // TODO: Generate prefunded accounts
@@ -17,11 +17,4 @@ pub fn handle(project_path: &Path) -> Result<()> {
     );
     diem_node::start(&node_config, None);
     Ok(())
-}
-
-fn read_config(project_path: &Path) -> Result<Config> {
-    let config_string =
-        fs::read_to_string(project_path.join("Shuffle").with_extension("toml")).unwrap();
-    let read_config: Config = toml::from_str(config_string.as_str())?;
-    Ok(read_config)
 }
