+++ shuffle/cli/src/shared.rs
@@ -1,0 +1,21 @@
+// Copyright (c) The Diem Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+use anyhow::Result;
+use diem_types::account_address::AccountAddress;
+use serde::{Deserialize, Serialize};
+use std::{collections::BTreeMap, fs, path::Path};
+
+#[derive(Debug, Serialize, Deserialize, PartialEq)]
+#[serde(rename_all = "kebab-case")]
+pub struct Config {
+    pub(crate) blockchain: String,
+    pub(crate) named_addresses: BTreeMap<String, AccountAddress>,
+}
+
+pub fn read_config(project_path: &Path) -> Result<Config> {
+    let config_string =
+        fs::read_to_string(project_path.join("Shuffle").with_extension("toml")).unwrap();
+    let read_config: Config = toml::from_str(config_string.as_str())?;
+    Ok(read_config)
+}
